FinW_Face_Recognition_System/requirements.txt
python-dotenv>=1.0.0
ultralytics>=8.0.0
opencv-python
numpy>=1.24.0
scikit-learn>=1.3.0
python-telegram-bot>=20.0
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0
torch>=2.0.0
torchvision>=0.15.0
matplotlib>=3.3.0
fastapi
opencv-python
pandas
python-dotenv
telegram
uvicorn